✨ (fuel-price): Fix bill creation and price calculation with current VAT (1.19)

+++ src/main/java/de/codeflowwizardry/carledger/rest/records/BillInputPojo.java
@@ -3,7 +3,7 @@ package de.codeflowwizardry.carledger.rest.records;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record BillInputPojo(LocalDate day, BigDecimal distance, Double unit, Double pricePerUnit, Double estimate)
+public record BillInputPojo(LocalDate day, BigDecimal distance, BigDecimal unit, BigDecimal pricePerUnit,
+		BigDecimal estimate)
 {
-
 }
